Update README.md
In "Beauty and the Beast," Belle, this smart and kind gal, gets stuck in a magical castle because her dad's in a jam. At first, she's freaked out by the Beast who lives there, but then she sees he's not so bad and befriends his enchanted furniture friends. They fall in love, break the curse with some true love magic, and bam, Beast turns back into a prince, and the castle's back to normal.